WEEK13/miniprojet2/mini-mcp-research-notebook-part2-FULL/pyproject.toml
[project]
name = "mini-mcp-research-notebook"
version = "0.2.0"
description = "Research Notebook: arXiv + Filesystem + Citation Cleaner + Scholar Plus via MCP"
requires-python = ">=3.10"
dependencies = [
  "mcp",
  "python-dotenv",
  "httpx",
  "pydantic",
  "streamlit",
  "groq",
  "ollama",
  "tenacity"
]
